requirements_dev.txt
pytest==3.0.4
pytest_flask==0.10.0
